rename KafkaError to APIError to avoid package name conflict
